+++ CHANGELOG.md
@@ -1,6 +1,18 @@
-# 0.1.0
+# 0.0.5
 
-Adds NlpPure::NGram::Default module.
+Fixed bug in `NlpPure::Segmenting::DefaultWord` where leading ellipses could produce extra segmented words.
+
+# 0.0.4
+
+Fixed bug in `NlpPure::Segmenting::DefaultWord` where ellipses without spaces would not segment.
+
+# 0.0.3
+
+Fixed bug in `NlpPure::Segmenting::DefaultWord` where double hyphens and spaced dashes would segment as empty words.
+
+# 0.0.2
+
+Added `NlpPure::Segmenting::DefaultWord` module for segmenting text into words.
 
 # 0.0.1